// src/shared/types/generated/application/CandidateSide.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CandidateOccurrence } from "./CandidateOccurrence";
import type { FileQuality } from "./FileQuality";

export type CandidateSide = { file: FileQuality, occurrences: Array<CandidateOccurrence>, };